Re-home the crypto pre-trans handler
