+++ hedera-node/hedera-app/src/test/java/com/hedera/node/app/workflows/dispatcher/DispatcherTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 Hedera Hashgraph, LLC
+ * Copyright (C) 2022-2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -41,6 +41,7 @@ import com.hedera.node.app.service.network.impl.handlers.UncheckedSubmitHandler;
 import com.hedera.node.app.service.schedule.impl.handlers.ScheduleCreateHandler;
 import com.hedera.node.app.service.schedule.impl.handlers.ScheduleDeleteHandler;
 import com.hedera.node.app.service.schedule.impl.handlers.ScheduleSignHandler;
+import com.hedera.node.app.service.token.CryptoSignatureWaivers;
 import com.hedera.node.app.service.token.impl.handlers.CryptoAddLiveHashHandler;
 import com.hedera.node.app.service.token.impl.handlers.CryptoApproveAllowanceHandler;
 import com.hedera.node.app.service.token.impl.handlers.CryptoCreateHandler;
@@ -139,6 +140,7 @@ class DispatcherTest {
     @Mock private TokenUnpauseHandler tokenUnpauseHandler;
 
     @Mock private UtilPrngHandler utilPrngHandler;
+    @Mock private CryptoSignatureWaivers cryptoSignatureWaivers;
 
     private Handlers handlers;
     private Dispatcher dispatcher;
@@ -194,15 +196,15 @@ class DispatcherTest {
                         tokenUnpauseHandler,
                         utilPrngHandler);
 
-        dispatcher = new Dispatcher(handlers, storeCache);
+        dispatcher = new Dispatcher(handlers, storeCache, cryptoSignatureWaivers);
     }
 
     @SuppressWarnings("ConstantConditions")
     @Test
     void testConstructorWithIllegalParameters() {
-        assertThatThrownBy(() -> new Dispatcher(null, storeCache))
+        assertThatThrownBy(() -> new Dispatcher(null, storeCache, cryptoSignatureWaivers))
                 .isInstanceOf(NullPointerException.class);
-        assertThatThrownBy(() -> new Dispatcher(handlers, null))
+        assertThatThrownBy(() -> new Dispatcher(handlers, null, cryptoSignatureWaivers))
                 .isInstanceOf(NullPointerException.class);
     }
 
@@ -335,7 +337,9 @@ class DispatcherTest {
                                         CryptoCreateTransactionBody.getDefaultInstance())
                                 .build(),
                         (Consumer<Handlers>)
-                                h -> verify(h.cryptoCreateHandler()).preHandle(any(), any())),
+                                h ->
+                                        verify(h.cryptoCreateHandler())
+                                                .preHandle(any(), any(), any())),
                 Arguments.of(
                         TransactionBody.newBuilder()
                                 .setCryptoUpdateAccount(
@@ -344,7 +348,7 @@ class DispatcherTest {
                         (Consumer<Handlers>)
                                 h ->
                                         verify(h.cryptoUpdateHandler())
-                                                .preHandle(any(), any(), any())),
+                                                .preHandle(any(), any(), any(), any())),
                 Arguments.of(
                         TransactionBody.newBuilder()
                                 .setCryptoTransfer(
@@ -444,13 +448,17 @@ class DispatcherTest {
                                         ScheduleCreateTransactionBody.getDefaultInstance())
                                 .build(),
                         (Consumer<Handlers>)
-                                h -> verify(h.scheduleCreateHandler()).preHandle(any(), any())),
+                                h ->
+                                        verify(h.scheduleCreateHandler())
+                                                .preHandle(any(), any(), any(), any())),
                 Arguments.of(
                         TransactionBody.newBuilder()
                                 .setScheduleSign(ScheduleSignTransactionBody.getDefaultInstance())
                                 .build(),
                         (Consumer<Handlers>)
-                                h -> verify(h.scheduleSignHandler()).preHandle(any(), any())),
+                                h ->
+                                        verify(h.scheduleSignHandler())
+                                                .preHandle(any(), any(), any(), any(), any())),
                 Arguments.of(
                         TransactionBody.newBuilder()
                                 .setScheduleDelete(
