+++ hedera-node/hedera-schedule-service-impl/src/main/java/com/hedera/node/app/service/schedule/impl/handlers/ScheduleCreateHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 Hedera Hashgraph, LLC
+ * Copyright (C) 2022-2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,14 @@
  */
 package com.hedera.node.app.service.schedule.impl.handlers;
 
+import static com.hedera.node.app.service.mono.Utils.asHederaKey;
+import static com.hedera.node.app.service.mono.utils.MiscUtils.asOrdinary;
+import static com.hedera.node.app.service.schedule.impl.handlers.ScheduleHandlerUtils.preHandleScheduledTxn;
+
+import com.hedera.node.app.spi.AccountKeyLookup;
+import com.hedera.node.app.spi.PreHandleDispatcher;
+import com.hedera.node.app.spi.meta.ScheduleSigTransactionMetadataBuilder;
+import com.hedera.node.app.spi.meta.ScheduleTransactionMetadata;
 import com.hedera.node.app.spi.meta.TransactionMetadata;
 import com.hedera.node.app.spi.workflows.TransactionHandler;
 import com.hederahashgraph.api.proto.java.AccountID;
@@ -37,15 +45,51 @@ public class ScheduleCreateHandler implements TransactionHandler {
      * <p>Please note: the method signature is just a placeholder which is most likely going to
      * change.
      *
-     * @param txBody the {@link TransactionBody} with the transaction data
+     * @param txn the {@link TransactionBody} with the transaction data
      * @param payer the {@link AccountID} of the payer
+     * @param keyLookup the {@link AccountKeyLookup} to use for key resolution
+     * @param dispatcher the {@link PreHandleDispatcher} that can be used to pre-handle the inner
+     *     txn
      * @return the {@link TransactionMetadata} with all information that needs to be passed to
      *     {@link #handle(TransactionMetadata)}
      * @throws NullPointerException if one of the arguments is {@code null}
      */
-    public TransactionMetadata preHandle(
-            @NonNull final TransactionBody txBody, @NonNull final AccountID payer) {
-        throw new UnsupportedOperationException("Not implemented");
+    public ScheduleTransactionMetadata preHandle(
+            @NonNull final TransactionBody txn,
+            @NonNull final AccountID payer,
+            @NonNull final AccountKeyLookup keyLookup,
+            @NonNull final PreHandleDispatcher dispatcher) {
+        final var op = txn.getScheduleCreate();
+        final var meta =
+                new ScheduleSigTransactionMetadataBuilder(keyLookup)
+                        .txnBody(txn)
+                        .payerKeyFor(payer);
+
+        if (op.hasAdminKey()) {
+            final var key = asHederaKey(op.getAdminKey());
+            key.ifPresent(meta::addToReqNonPayerKeys);
+        }
+
+        final var scheduledTxn =
+                asOrdinary(op.getScheduledTransactionBody(), txn.getTransactionID());
+
+        /* We need to always add the custom payer to the sig requirements even if it equals the to level transaction
+        payer. It is still part of the "other" parties, and we need to know to store it's key with the
+        schedule in all cases. This fixes a case where the ScheduleCreate payer and the custom payer are
+        the same payer, which would cause the custom payers signature to not get stored and then a ScheduleSign
+        would not execute the transaction without and extra signature from the custom payer.*/
+        final var payerForNested =
+                op.hasPayerAccountID()
+                        ? op.getPayerAccountID()
+                        : txn.getTransactionID().getAccountID();
+
+        // FUTURE: Once we allow schedule transactions to be scheduled inside, we need a check here
+        // to see
+        // if provided payer is same as payer in the inner transaction.
+
+        final var innerMeta = preHandleScheduledTxn(scheduledTxn, payerForNested, dispatcher);
+        meta.scheduledMeta(innerMeta);
+        return meta.build();
     }
 
     /**
