+++ hedera-node/hedera-schedule-service-impl/src/main/java/com/hedera/node/app/service/schedule/impl/handlers/ScheduleSignHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 Hedera Hashgraph, LLC
+ * Copyright (C) 2022-2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,15 @@
  */
 package com.hedera.node.app.service.schedule.impl.handlers;
 
+import static com.hedera.node.app.service.schedule.impl.handlers.ScheduleHandlerUtils.preHandleScheduledTxn;
+import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_SCHEDULE_ID;
+
+import com.hedera.node.app.service.schedule.impl.ReadableScheduleStore;
+import com.hedera.node.app.spi.AccountKeyLookup;
+import com.hedera.node.app.spi.PreHandleDispatcher;
+import com.hedera.node.app.spi.meta.InvalidTransactionMetadata;
+import com.hedera.node.app.spi.meta.ScheduleSigTransactionMetadataBuilder;
+import com.hedera.node.app.spi.meta.ScheduleTransactionMetadata;
 import com.hedera.node.app.spi.meta.TransactionMetadata;
 import com.hedera.node.app.spi.workflows.TransactionHandler;
 import com.hederahashgraph.api.proto.java.AccountID;
@@ -37,15 +46,43 @@ public class ScheduleSignHandler implements TransactionHandler {
      * <p>Please note: the method signature is just a placeholder which is most likely going to
      * change.
      *
-     * @param txBody the {@link TransactionBody} with the transaction data
+     * @param txn the {@link TransactionBody} with the transaction data
      * @param payer the {@link AccountID} of the payer
      * @return the {@link TransactionMetadata} with all information that needs to be passed to
      *     {@link #handle(TransactionMetadata)}
+     * @param keyLookup the {@link AccountKeyLookup} to use for key resolution
+     * @param scheduleStore the {@link ReadableScheduleStore} to use for schedule resolution
+     * @param dispatcher the {@link PreHandleDispatcher} that can be used to pre-handle the inner
+     *     txn
      * @throws NullPointerException if one of the arguments is {@code null}
      */
-    public TransactionMetadata preHandle(
-            @NonNull final TransactionBody txBody, @NonNull final AccountID payer) {
-        throw new UnsupportedOperationException("Not implemented");
+    public ScheduleTransactionMetadata preHandle(
+            @NonNull final TransactionBody txn,
+            @NonNull final AccountID payer,
+            @NonNull final AccountKeyLookup keyLookup,
+            @NonNull final ReadableScheduleStore scheduleStore,
+            @NonNull final PreHandleDispatcher dispatcher) {
+        final var op = txn.getScheduleSign();
+        final var id = op.getScheduleID();
+
+        final var scheduleLookupResult = scheduleStore.get(id);
+        if (scheduleLookupResult.isEmpty()) {
+            return new InvalidTransactionMetadata(txn, payer, INVALID_SCHEDULE_ID);
+        }
+
+        final var meta =
+                new ScheduleSigTransactionMetadataBuilder(keyLookup)
+                        .txnBody(txn)
+                        .payerKeyFor(payer);
+
+        final var scheduledTxn = scheduleLookupResult.get().scheduledTxn();
+        final var optionalPayer = scheduleLookupResult.get().designatedPayer();
+        final var payerForNested =
+                optionalPayer.orElse(scheduledTxn.getTransactionID().getAccountID());
+
+        final var innerMeta = preHandleScheduledTxn(scheduledTxn, payerForNested, dispatcher);
+        meta.scheduledMeta(innerMeta);
+        return meta.build();
     }
 
     /**
