+++ hedera-node/hedera-schedule-service-impl/src/main/java/module-info.java
@@ -8,8 +8,11 @@ module com.hedera.node.app.service.schedule.impl {
 
     exports com.hedera.node.app.service.schedule.impl to
             com.hedera.node.app.service.schedule.impl.test,
-            com.hedera.node.app.service.scheduled.impl.test;
-    exports com.hedera.node.app.service.schedule.impl.handlers;
+            com.hedera.node.app.service.scheduled.impl.test,
+            com.hedera.node.app;
+    exports com.hedera.node.app.service.schedule.impl.handlers to
+            com.hedera.node.app.service.schedule.impl.test,
+            com.hedera.node.app;
 
     provides com.hedera.node.app.service.schedule.ScheduleService with
             ScheduleServiceImpl;
